bump libs

+++ src/Maw.Domain.Models/Search/SearchResults.cs
@@ -3,6 +3,6 @@ namespace Maw.Domain.Search;
 public class SearchResults<T>
 {
     public IEnumerable<T> Results { get; set; } = new List<T>();
-    public int TotalFound { get; set; }
-    public int StartIndex { get; set; }
+    public long TotalFound { get; set; }
+    public long StartIndex { get; set; }
 }
